feat: 日志列表智能分组与交互优化
- 新增 LogListItem sealed class (Item/Group) 多 ViewType 数据类型
- 新增 LogGrouper 分组算法 (content+tag+level 相同判定, 默认100条/组)
- 新增 ScrollFabIcon 枚举控制 FAB 双向导航
- 新增 GroupViewHolder 分组头部 ViewHolder + item_group_log.xml 布局
- 新增方向箭头图标 ic_arrow_drop_up/down
- BaseListAdapter 支持多 ViewType (abstract→open 保持向后兼容)
- LogsAdapter 多 ViewType + setHasStableIds + 自定义 DiffCallback
- LogsState 新增 groupExpandStates/isAtBottom, LogsViewState 新增 scrollFabIcon
- LogsCommand 新增 GroupToggled/ExpandAllGroups/CollapseAllGroups/ScrollStateChanged
- LogsReducer 处理新 Command, LogsViewStateMapper 集成 LogGrouper
- LogViewHolder 单击复制 + 透明度高亮动画
- LogsFragment FAB 双向导航 + 滚动监听 + RecyclerView 性能配置
- 修复数据库模块 schema 文件缺失 (18→19,19→20,20→21 合并为 18→21)

+++ feature/preferences/presentation/src/main/kotlin/com/f0x1d/logfox/feature/preferences/presentation/menu/PreferencesMenuViewState.kt
@@ -2,6 +2,6 @@ package com.f0x1d.logfox.feature.preferences.presentation.menu
 
 internal data class PreferencesMenuViewState(
     val versionName: String,
-    val versionCode: Int,
+    val versionCode: Long,
     val isDebug: Boolean,
 )
